fix: rename util.rs to cleaner.rs and update lib.rs
